refac: creating Overview card as component
